883: build: update nixpkgs and deps r=tiagolobocastro a=tiagolobocastro

    test(bdd): update python packages
    

---

    fix(resize): atomically check for the required size
    
    Ensures races don't lead into volume resize failures.
    

---

    build: specify patches for e2fsprogs
    
    This is required because the upstream nixpkgs applies patches which don't
    apply on the old version we're building..
    

---

    build!: update kube and related crates
    
    Brings back kube-proxy crate to this repo, for easier managing of code
    changes related to the openapi service.
    Updates kube to 0.94.2 and k8s-openapi to 0.22.0 with K8s v1_24
    

---

    build: upgrade tracing, opentel, tonic, prost, rustls, tls and hyper
    
    Upgrade all aforementioned crates and some related crates.
    

---

    build: update most crate dependencies
    
    Updated all except some which are a bit more problematic:
    hyper
    rustls
    k8s-openapi
    kube
    hyper-rustls
    hyper-tls
    opentelemetry-otlp
    h2
    http
    actix-web
    actix-web-opentelemetry
    opentelemetry_sdk
    tracing-opentelemetry
    opentelemetry
    opentelemetry-semantic-conventions
    opentelemetry-http
    opentelemetry-jaeger
    http-body
    tower
    tower-http
    rustls-pemfile
    tonic
    prost
    tonic-build
    prost-build
    prost-derive
    prost-types
    

---

    test: temporarily use previous openapi-generator-cli
    
    The newer versions have removed the python-prior generator and so
    upgrading is a considerable amount of work.
    Let's pin the nixpkgs with the openeapi-generator-cli and we'll
    tackle this as a separate task.
    

---

    build: update nixpkgs and deps

+++ tests/bdd/common/operations.py
@@ -138,6 +138,9 @@ class Cluster(object):
         for node in ApiClient.nodes_api().get_nodes():
             wait_node_online(node.id)
 
+        if not Cluster.fixture_cleanup():
+            return
+
         # ensure snapshots are all delete
         if snapshots:
             Snapshot.delete_all()
